fix: code review patches — endgame safety, session cleanup, input validation
Critical:
- Battle.ts: guard endgame() with battle.winner === null to prevent double renown on concurrent kills (C-1)
- Battle.ts: clear session.battle_id on /exit so re-queuing and queue notifications work after a match (C-2)
- app.ts: wrap /debug/* endpoints in NODE_ENV !== production guard (C-3)

Medium:
- Battle.ts: validate killedparty/killerparty against known aliveUnits keys before processing kill state (M-4)
- queue.ts: use session.account_id (32-bit) instead of session.user_id in QueueItem and duplicate-queue guard (M-1)
- roster.ts: validate new_unit_name type and length in /unit/hire, matching rename/promote validation (M-2)
- discord.ts: reject login instead of warn-and-continue when Discord Snowflake ID loses precision (M-6)
- game.ts: cache lboard.json at module load instead of reading on every /leaderboards request (M-7)
- download.ts: pre-compute factions.tar.gz checksum at startup instead of hashing on every /checksum request (M-7)

Also includes pre-existing uncommitted work:
- src/index.ts + src/app.ts: extract Express app from index.ts into app.ts for cleaner separation
- test-2p-match.bat: use Write-Output, add timeout before poll, tighten error messages

GitHub issues opened for deferred items: #37 (M-3 promote mutation order), #38 (M-5 app._router internals), #32-36 (L-1 through L-5)
